Add files via upload
The TCD1304 is a high sensitivity 3648 pixel linear CCD. It's used in some Ocean Optics spectrometers, but because of the difficulty of driving CCDs it has not found widespread use in DIY projects. This project describes a low cost linear CCD (Linear CCD Module SPI Serial Port USB TCD1304 Single Chip CCD Precision Optical Measurement Customization ) module based on the TCD1304. The TCD1304 is connected to an STM32 microcontroller, which takes care of the time-critical driving and reading of the CCD. 
The ASCII code format, where a can be a, B, C, D, F, represents a variety of output modes, in which a, B output with coordinates, such as 253-0x3AE, the 253 representation in the front, and the 12 digits behind the AD sampling value. C, D mode is a continuous output AD sampling value (requiring the customer to make a single request for convenient and fast transmission of data), without output coordinates, so that it can be output faster. The f mode is continuous sampling without output value. The output waveforms of sensors can be monitored by an oscilloscope. The baud rate is 115200, N, 8,1. 
0080 represents integral time, which ranges from 0005-0256 to higher. The larger the value is, the longer the exposure time is, and the smaller the output value. 
TCD1304 output, the stronger the light, the smaller the value. Because the sensor pixel size 8*200um is wider, so it is more sensitive, very easy to be affected by ambient light, easy to saturation, and needs attention when testing. 
At present, all serial products are sold, and SPI is custom-made for customers. 
  
A brief introduction to 2:SPI 
This is an image acquisition module based on linear CCD. Linear array CCD is a kind of photoelectric detection sensor, which is used in the field of light signal analysis, image scanning, machine vision and so on. The MPS-CCD Driver module integrates linear CCD sensor, drive acquisition unit and SPI bus communication interface, which can be used as a component of photoelectric detection system. When using the module to design the photoelectric detection system, it is unnecessary to pay attention to the hardware circuit design of the signal acquisition part, which greatly reduces the difficulty of the system design. 
The Toshiba Co's high-sensitivity linear array CCD contains 3048 usable pixels, and the pixel spacing is 8 microns. The exposure time of CCD is adjustable with 254 stages, which can meet the requirements of various low light level testing occasions. The module uses 12 bit high resolution ADC to sample the sensor, and the sampling frame rate can reach more than 5 frames per second. It has the characteristics of high speed and high accuracy. 
MPS-CCD Driver uses standard SPI bus to communicate with external devices, and it can be applied to various embedded processors such as MCU, ARM and FPGA. The module is used as SPI slave device, and the external controller can control the module through SPI communication for image acquisition and data transmission. Communication protocol is simple and easy to use, highly reliable, and easy to integrate with other applications. Combined with the image analysis and processing program inside the controller, the function design of a set of high-performance photoelectric analysis instrument can be easily realized.